08-evals-for-ai-models/requirements.txt
openai>=1.0.0
pandas>=2.0.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
ragas>=0.1.0
